supersub_backend/requirements.txt
fastapi[standard]
sqlmodel
psycopg[binary]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
redis